requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
dateparser==1.2.1
deepseek==1.0.0
distro==1.9.0
fonttools==4.56.0
fpdf==1.7.2
frozendict==2.4.6
frozenlist==1.5.0
google-auth==2.38.0
google-genai==1.4.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
joblib==1.4.2
kiwisolver==1.4.8
lxml==5.3.1
matplotlib==3.10.0
multidict==6.1.0
multitasking==0.0.11
nltk==3.9.1
numpy==2.2.3
oauthlib==3.2.2
openai==1.64.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pandas_ta==0.3.14b0
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.6
propcache==0.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
schedule==1.2.2
selenium==4.29.0
setuptools==75.8.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.38
textblob==0.19.0
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
tweepy==4.15.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
undetected-chromedriver==3.5.5
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==14.2
wsproto==1.2.0
yarl==1.18.3
yfinance==0.2.54
